+++ src/IOWrapper/Pangolin/PangolinDSOViewer.cpp
@@ -419,7 +419,7 @@ void PangolinDSOViewer::drawConstraints()
 
 
 
-void PangolinDSOViewer::publishGraph(const std::map<long,Eigen::Vector2i> &connectivity)
+void PangolinDSOViewer::publishGraph(const std::map<uintptr_t,Eigen::Vector2i> &connectivity)
 {
     if(!setting_render_display3D) return;
     if(disableAllDisplay) return;
@@ -428,7 +428,7 @@ void PangolinDSOViewer::publishGraph(const std::map<long,Eigen::Vector2i> &conne
 	connections.resize(connectivity.size()/2);
 	int runningID=0;
 	int totalActFwd=0, totalActBwd=0, totalMargFwd=0, totalMargBwd=0;
-	for(std::pair<long,Eigen::Vector2i> p : connectivity)
+	for(std::pair<uintptr_t,Eigen::Vector2i> p : connectivity)
 	{
 		int host = (int)(p.first >> 32);
 		int target = (int)(p.first & (long)0xFFFFFFFF);
@@ -449,7 +449,7 @@ void PangolinDSOViewer::publishGraph(const std::map<long,Eigen::Vector2i> &conne
 		totalActFwd += p.second[0];
 		totalMargFwd += p.second[1];
 
-		long inverseKey = (((long)target) << 32) + ((long)host);
+		uintptr_t inverseKey = (((uintptr_t)target) << 32) + ((uintptr_t)host);
 		Eigen::Vector2i st = connectivity.at(inverseKey);
 		connections[runningID].bwdAct = st[0];
 		connections[runningID].bwdMarg = st[1];
